refactor(server): trim unmapped event repair
