+++ apps/server/src/orchestration/providerRuntimeActivityProjection.ts
@@ -9,13 +9,16 @@ import {
 } from "@synara/contracts";
 import { nonEmptyTrimmed } from "@synara/shared/text";
 
-import { activityDataField, boundActivityData, stringifyJsonLike } from "../activityData.ts";
 import {
   sanitizeUnmappedProviderData,
   sanitizeUnmappedProviderDetail,
 } from "../provider/unmappedProviderEvents.ts";
 
+const MAX_ACTIVITY_DATA_JSON_CHARS = 16_000;
 const MAX_ACTIVITY_DATA_STRING_CHARS = 2_000;
+const MAX_ACTIVITY_DATA_ARRAY_ITEMS = 24;
+const MAX_ACTIVITY_DATA_OBJECT_KEYS = 64;
+const ACTIVITY_DATA_TRUNCATION_MARKER = "__synaraTruncated";
 
 type ActivityPayload = OrchestrationThreadActivity["payload"];
 
@@ -106,6 +109,194 @@ function truncateDetail(value: string, limit = 180): string {
   return value.length > limit ? `${value.slice(0, limit - 3)}...` : value;
 }
 
+function stringifyJsonLike(value: unknown): string {
+  const seen = new WeakSet<object>();
+  return (
+    JSON.stringify(value, (_key, entry) => {
+      if (typeof entry === "bigint") {
+        return entry.toString();
+      }
+      if (typeof entry === "function" || typeof entry === "symbol") {
+        return undefined;
+      }
+      if (entry && typeof entry === "object") {
+        if (seen.has(entry)) {
+          return "[Circular]";
+        }
+        seen.add(entry);
+      }
+      return entry;
+    }) ?? "null"
+  );
+}
+
+function truncateJsonString(value: string, limit: number): string {
+  return value.length > limit ? `${value.slice(0, Math.max(0, limit - 15))}... [truncated]` : value;
+}
+
+function isJsonObject(value: unknown): value is Record<string, unknown> {
+  return value !== null && typeof value === "object" && !Array.isArray(value);
+}
+
+function activityPayloadKeyRank(key: string): number {
+  const ranks: Record<string, number> = {
+    itemType: 0,
+    status: 1,
+    title: 2,
+    detail: 3,
+    toolName: 4,
+    tool: 5,
+    toolCallId: 6,
+    callID: 7,
+    callId: 8,
+    command: 9,
+    cmd: 10,
+    input: 11,
+    rawInput: 12,
+    arguments: 13,
+    args: 14,
+    params: 15,
+    item: 16,
+    result: 17,
+    rawOutput: 18,
+    output: 19,
+    data: 20,
+    commandActions: 21,
+    files: 22,
+    changes: 23,
+    path: 24,
+    file: 25,
+    filePath: 26,
+    stdout: 27,
+    stderr: 28,
+    content: 29,
+    totalFiles: 30,
+    truncated: 31,
+  };
+  return ranks[key] ?? 100;
+}
+
+function truncateJsonValue(
+  value: unknown,
+  options: {
+    readonly stringLimit: number;
+    readonly arrayItems: number;
+    readonly objectKeys: number;
+    readonly depth: number;
+    readonly seen?: WeakSet<object>;
+  },
+): unknown {
+  if (value === null || typeof value === "number" || typeof value === "boolean") {
+    return value;
+  }
+  if (typeof value === "string") {
+    return truncateJsonString(value, options.stringLimit);
+  }
+  if (typeof value === "bigint") {
+    return value.toString();
+  }
+  if (typeof value === "function" || typeof value === "symbol" || value === undefined) {
+    return null;
+  }
+  const seen = options.seen ?? new WeakSet<object>();
+  if (value && typeof value === "object") {
+    if (seen.has(value)) {
+      return "[Circular]";
+    }
+    seen.add(value);
+  }
+  if (options.depth <= 0) {
+    return isJsonObject(value) || Array.isArray(value)
+      ? {
+          [ACTIVITY_DATA_TRUNCATION_MARKER]: true,
+        }
+      : String(value);
+  }
+  if (Array.isArray(value)) {
+    const retained = value
+      .slice(0, options.arrayItems)
+      .map((entry) => truncateJsonValue(entry, { ...options, depth: options.depth - 1 }));
+    if (value.length > options.arrayItems) {
+      retained.push({
+        [ACTIVITY_DATA_TRUNCATION_MARKER]: true,
+        omittedItems: value.length - options.arrayItems,
+      });
+    }
+    return retained;
+  }
+  if (!isJsonObject(value)) {
+    return String(value);
+  }
+
+  const entries = Object.entries(value)
+    .filter(
+      ([, entry]) =>
+        entry !== undefined && typeof entry !== "function" && typeof entry !== "symbol",
+    )
+    .toSorted((left, right) => {
+      const byRank = activityPayloadKeyRank(left[0]) - activityPayloadKeyRank(right[0]);
+      return byRank !== 0 ? byRank : left[0].localeCompare(right[0]);
+    });
+  const retainedEntries = entries.slice(0, options.objectKeys);
+  const result: Record<string, unknown> = {};
+  for (const [key, entry] of retainedEntries) {
+    result[key] = truncateJsonValue(entry, { ...options, depth: options.depth - 1 });
+  }
+  if (entries.length > options.objectKeys) {
+    result[ACTIVITY_DATA_TRUNCATION_MARKER] = true;
+    result.omittedKeys = entries.length - options.objectKeys;
+  }
+  return result;
+}
+
+function boundActivityData(value: unknown): unknown {
+  const serialized = stringifyJsonLike(value);
+  if (serialized.length <= MAX_ACTIVITY_DATA_JSON_CHARS) {
+    return JSON.parse(serialized);
+  }
+
+  const withTruncationMetadata = (bounded: unknown): Record<string, unknown> => {
+    const metadata = {
+      [ACTIVITY_DATA_TRUNCATION_MARKER]: true,
+      originalJsonChars: serialized.length,
+    };
+    return isJsonObject(bounded) ? { ...bounded, ...metadata } : { ...metadata, value: bounded };
+  };
+  const hardFallback = (): Record<string, unknown> => ({
+    [ACTIVITY_DATA_TRUNCATION_MARKER]: true,
+    originalJsonChars: serialized.length,
+    preview: truncateJsonString(serialized, MAX_ACTIVITY_DATA_STRING_CHARS),
+  });
+
+  const compact = truncateJsonValue(value, {
+    stringLimit: MAX_ACTIVITY_DATA_STRING_CHARS,
+    arrayItems: MAX_ACTIVITY_DATA_ARRAY_ITEMS,
+    objectKeys: MAX_ACTIVITY_DATA_OBJECT_KEYS,
+    depth: 6,
+  });
+  const compactWithMetadata = withTruncationMetadata(compact);
+  if (stringifyJsonLike(compactWithMetadata).length <= MAX_ACTIVITY_DATA_JSON_CHARS) {
+    return compactWithMetadata;
+  }
+
+  const bounded = withTruncationMetadata(
+    truncateJsonValue(value, {
+      stringLimit: 800,
+      arrayItems: 12,
+      objectKeys: 32,
+      depth: 4,
+    }),
+  );
+  return stringifyJsonLike(bounded).length <= MAX_ACTIVITY_DATA_JSON_CHARS
+    ? bounded
+    : hardFallback();
+}
+
+// Tool payloads power the timeline, but they must stay small enough for snapshots.
+function activityDataField(data: unknown): { readonly data?: unknown } {
+  return data === undefined ? {} : { data: boundActivityData(data) };
+}
+
 // Keep MCP progress payloads available to the web timeline so it can render the specific tool call.
 function buildToolProgressActivityPayload(
   event: Extract<ProviderRuntimeEvent, { type: "tool.progress" }>,
@@ -132,10 +323,6 @@ export function readableReasoningDetail(value: string | undefined): string | und
 
 function asString(value: unknown): string | undefined {
   return typeof value === "string" ? value : undefined;
-}
-
-function isJsonObject(value: unknown): value is Record<string, unknown> {
-  return value !== null && typeof value === "object" && !Array.isArray(value);
 }
 
 function asObject(value: unknown): Record<string, unknown> | undefined {
@@ -944,8 +1131,6 @@ export function projectProviderRuntimeActivities(
       }
       const detail = asString(payload?.detail);
       const rawData = payload?.data;
-      // Bound and redact again at the durable activity boundary. This also
-      // protects replay of older journal rows written before ingress hardening.
       return [
         {
           id: event.eventId,
@@ -956,9 +1141,7 @@ export function projectProviderRuntimeActivities(
           payload: toActivityPayload({
             nativeEventType: nativeType,
             ...(detail ? { detail: sanitizeUnmappedProviderDetail(detail) } : {}),
-            ...(rawData === undefined
-              ? {}
-              : activityDataField(sanitizeUnmappedProviderData(rawData))),
+            ...(rawData !== undefined ? { data: sanitizeUnmappedProviderData(rawData) } : {}),
           }),
           turnId: toTurnId(event.turnId) ?? null,
           ...maybeSequence,
