+++ apps/server/src/provider/Layers/CodexAdapter.test.ts
@@ -95,7 +95,6 @@ class FakeCodexManager extends CodexAppServerManager {
   );
 
   public stopAllImpl = vi.fn(() => undefined);
-  public stopSessionImpl = vi.fn(async (_threadId: ThreadId): Promise<void> => undefined);
 
   override startSession(input: CodexAppServerStartSessionInput): Promise<ProviderSession> {
     return this.startSessionImpl(input);
@@ -141,9 +140,7 @@ class FakeCodexManager extends CodexAppServerManager {
     return this.respondToUserInputImpl(threadId, requestId, answers);
   }
 
-  override stopSession(threadId: ThreadId): Promise<void> {
-    return this.stopSessionImpl(threadId);
-  }
+  override async stopSession(_threadId: ThreadId): Promise<void> {}
 
   override listSessions(): ProviderSession[] {
     return [];
@@ -1561,70 +1558,63 @@ lifecycleLayer("CodexAdapterLive lifecycle", (it) => {
     }),
   );
 
-  it.effect("surfaces unmapped native events with bounded redacted diagnostics", () =>
-    Effect.gen(function* () {
-      const adapter = yield* CodexAdapter;
-      const firstEventFiber = yield* Stream.runHead(adapter.streamEvents).pipe(Effect.forkChild);
+  it.effect(
+    "surfaces previously-unmapped native events with bounded redacted diagnostics instead of raw payloads",
+    () =>
+      Effect.gen(function* () {
+        const adapter = yield* CodexAdapter;
+        const firstEventFiber = yield* Stream.runHead(adapter.streamEvents).pipe(Effect.forkChild);
 
-      // `item/agentMessage/completed` has no explicit mapping (only the
-      // `item/agentMessage/delta` stream does); before the passthrough
-      // fallback this event produced no runtime event at all.
-      lifecycleManager.emit("event", {
-        id: asEventId("evt-unmapped-agent-message-completed"),
-        kind: "notification",
-        provider: "codex",
-        createdAt: new Date().toISOString(),
-        method: "item/agentMessage/completed",
-        threadId: asThreadId("thread-1"),
-        turnId: asTurnId("turn-1"),
-        itemId: asItemId("agent_message_9"),
-        payload: {
-          apiKey: "must-not-reach-the-runtime-journal",
-          msg: {
-            type: "item/agentMessage/completed",
-            item_id: "agent_message_9",
-            summary: "Finished the refactor",
+        // `item/agentMessage/completed` has no explicit mapping (only the
+        // `item/agentMessage/delta` stream does); before the passthrough
+        // fallback this event produced no runtime event at all.
+        lifecycleManager.emit("event", {
+          id: asEventId("evt-unmapped-agent-message-completed"),
+          kind: "notification",
+          provider: "codex",
+          createdAt: new Date().toISOString(),
+          method: "item/agentMessage/completed",
+          threadId: asThreadId("thread-1"),
+          turnId: asTurnId("turn-1"),
+          itemId: asItemId("agent_message_9"),
+          payload: {
+            apiKey: "must-not-reach-the-runtime-journal",
+            note: "Authorization: Bearer private-token",
+            msg: {
+              type: "item/agentMessage/completed",
+              item_id: "agent_message_9",
+              summary: "Finished the refactor",
+            },
+            output: "x".repeat(64_000),
           },
-        },
-      } satisfies ProviderEvent);
+        } satisfies ProviderEvent);
 
-      const firstEvent = yield* Fiber.join(firstEventFiber);
-      assert.equal(firstEvent._tag, "Some");
-      if (firstEvent._tag !== "Some") {
-        return;
-      }
-      assert.equal(firstEvent.value.type, "event.unmapped");
-      if (firstEvent.value.type !== "event.unmapped") {
-        return;
-      }
-      // Raw native type/label is carried as the title source.
-      assert.equal(firstEvent.value.payload.nativeType, "item/agentMessage/completed");
-      // Readable one-liner extracted before the diagnostic is sanitized.
-      assert.equal(firstEvent.value.payload.detail, "Finished the refactor");
-      // Secret-shaped fields are redacted before journal admission.
-      assert.deepEqual(firstEvent.value.payload.data, {
-        apiKey: "[REDACTED]",
-        msg: {
-          type: "item/agentMessage/completed",
-          item_id: "agent_message_9",
-          summary: "Finished the refactor",
-        },
-      });
-      assert.equal(
-        JSON.stringify(firstEvent.value).includes("must-not-reach-the-runtime-journal"),
-        false,
-      );
-      assert.deepEqual(firstEvent.value.raw?.payload, {
-        synaraSanitized: true,
-        reason: "unmapped provider payload retained in bounded event data",
-      });
-      // Provider refs still resolved from the raw event.
-      assert.equal(firstEvent.value.itemId, "agent_message_9");
-      assert.equal(firstEvent.value.providerRefs?.providerItemId, "agent_message_9");
-    }),
+        const firstEvent = yield* Fiber.join(firstEventFiber);
+        assert.equal(firstEvent._tag, "Some");
+        if (firstEvent._tag !== "Some") {
+          return;
+        }
+        assert.equal(firstEvent.value.type, "event.unmapped");
+        if (firstEvent.value.type !== "event.unmapped") {
+          return;
+        }
+        // Raw native type/label is carried as the title source.
+        assert.equal(firstEvent.value.payload.nativeType, "item/agentMessage/completed");
+        assert.equal(firstEvent.value.payload.detail, "Finished the refactor");
+        const serialized = JSON.stringify(firstEvent.value);
+        assert.equal(serialized.includes("must-not-reach-the-runtime-journal"), false);
+        assert.equal(serialized.includes("private-token"), false);
+        assert.ok(serialized.length < 17_000);
+        assert.deepEqual(firstEvent.value.raw?.payload, {
+          synaraSanitized: true,
+        });
+        // Provider refs still resolved from the raw event.
+        assert.equal(firstEvent.value.itemId, "agent_message_9");
+        assert.equal(firstEvent.value.providerRefs?.providerItemId, "agent_message_9");
+      }),
   );
 
-  it.effect("coalesces repeated unmapped burst events within one turn", () =>
+  it.effect("coalesces repeated unmapped burst events", () =>
     Effect.gen(function* () {
       const adapter = yield* CodexAdapter;
       const eventsFiber = yield* Stream.take(adapter.streamEvents, 2).pipe(
@@ -1654,41 +1644,6 @@ lifecycleLayer("CodexAdapterLive lifecycle", (it) => {
         ),
         ["item/future/outputDelta", "item/future/completed"],
       );
-    }),
-  );
-
-  it.effect("releases unmapped burst state when its session stops", () =>
-    Effect.gen(function* () {
-      const adapter = yield* CodexAdapter;
-      const eventsFiber = yield* Stream.take(adapter.streamEvents, 2).pipe(
-        Stream.runCollect,
-        Effect.forkChild,
-      );
-      const threadId = asThreadId("thread-unmapped-stopped");
-      const emitDelta = (id: string) =>
-        lifecycleManager.emit("event", {
-          id: asEventId(id),
-          kind: "notification",
-          provider: "codex",
-          createdAt: new Date().toISOString(),
-          method: "item/future/outputDelta",
-          threadId,
-          turnId: asTurnId("turn-abandoned"),
-          payload: { summary: id },
-        } satisfies ProviderEvent);
-
-      emitDelta("evt-unmapped-before-stop");
-      yield* adapter.stopSession(threadId);
-      emitDelta("evt-unmapped-after-stop");
-
-      const events = Array.from(yield* Fiber.join(eventsFiber).pipe(Effect.timeout("1 second")));
-      assert.deepEqual(
-        events.map((event) =>
-          event.type === "event.unmapped" ? event.payload.detail : event.type,
-        ),
-        ["evt-unmapped-before-stop", "evt-unmapped-after-stop"],
-      );
-      assert.deepEqual(lifecycleManager.stopSessionImpl.mock.lastCall, [threadId]);
     }),
   );
 });
