+++ apps/server/src/provider/providerRuntimeEventIngress.ts
@@ -1,10 +1,5 @@
 import type { ProviderRuntimeEvent } from "@synara/contracts";
 
-import {
-  sanitizeUnmappedProviderData,
-  sanitizeUnmappedProviderDetail,
-} from "./unmappedProviderEvents.ts";
-
 export const PROVIDER_RUNTIME_CALLBACK_BUFFER_MAX_BYTES = 32 * 1024 * 1024;
 export const PROVIDER_RUNTIME_CALLBACK_TERMINAL_RESERVE = 64;
 export const PROVIDER_RUNTIME_INGRESS_EVENT_MAX_BYTES = 512 * 1024;
@@ -21,34 +16,6 @@ export function providerRuntimeEventBytes(event: ProviderRuntimeEvent): number {
   }
 }
 
-function sanitizeUnmappedEventForIngress(event: ProviderRuntimeEvent): ProviderRuntimeEvent {
-  if (event.type !== "event.unmapped") {
-    return event;
-  }
-  const data =
-    event.payload.data === undefined ? undefined : sanitizeUnmappedProviderData(event.payload.data);
-  const detail = sanitizeUnmappedProviderDetail(event.payload.detail);
-  return {
-    ...event,
-    payload: {
-      ...event.payload,
-      ...(detail === undefined ? {} : { detail }),
-      ...(data === undefined ? {} : { data }),
-    },
-    ...(event.raw === undefined
-      ? {}
-      : {
-          raw: {
-            ...event.raw,
-            payload: {
-              synaraSanitized: true,
-              reason: "unmapped provider payload retained in bounded event data",
-            },
-          },
-        }),
-  };
-}
-
 /**
  * Raw provider payloads are diagnostic data. Compact them before the callback
  * ingress so one pathological native message cannot consume the whole budget.
@@ -56,22 +23,16 @@ function sanitizeUnmappedEventForIngress(event: ProviderRuntimeEvent): ProviderR
 export function compactProviderRuntimeEventForIngress(
   event: ProviderRuntimeEvent,
 ): ProviderRuntimeEvent {
-  const sanitizedEvent = sanitizeUnmappedEventForIngress(event);
-  const originalBytes = providerRuntimeEventBytes(sanitizedEvent);
-  if (
-    originalBytes <= PROVIDER_RUNTIME_INGRESS_EVENT_MAX_BYTES ||
-    sanitizedEvent.raw === undefined
-  ) {
-    return sanitizedEvent;
+  const originalBytes = providerRuntimeEventBytes(event);
+  if (originalBytes <= PROVIDER_RUNTIME_INGRESS_EVENT_MAX_BYTES || event.raw === undefined) {
+    return event;
   }
   return {
-    ...sanitizedEvent,
+    ...event,
     raw: {
-      source: sanitizedEvent.raw.source,
-      ...(sanitizedEvent.raw.method !== undefined ? { method: sanitizedEvent.raw.method } : {}),
-      ...(sanitizedEvent.raw.messageType !== undefined
-        ? { messageType: sanitizedEvent.raw.messageType }
-        : {}),
+      source: event.raw.source,
+      ...(event.raw.method !== undefined ? { method: event.raw.method } : {}),
+      ...(event.raw.messageType !== undefined ? { messageType: event.raw.messageType } : {}),
       payload: {
         synaraTruncated: true,
         reason: "provider runtime event exceeded the callback ingress size limit",
